Creates dev environment' scripts
At first I've attempted to build over Bitnami's Dokuwiki image,
but since it deploys Dokuwiki during the container initialization,
this strategy didn't work. So, I just wrote a bunch of shell scripts
that loads everything up:

  * start initializes the container, while uploading the longform
    template and setting it up as the current template.
  * stop turn the container off.
  * upload removes the container's template directory (if it's
    there), and upload it again.

So, it is just a matter of starting it, working, and uploading the
template changes every so often. And you can stop the container at
the end of your work session, if you wish. It's not as sexy as a
hot reload, but it works.

Note: the start script has a `sleep` instruction, that gives the
container some time to initialize. Depending on your hardware, you
may need to increase this sleep window to prevent the template
setup from running too fast.